Feature network restructure

+++ RCAIDE/Library/Methods/Energy/Distributors/Electrical_Bus/compute_bus_conditions.py
@@ -6,6 +6,9 @@
 # ----------------------------------------------------------------------------------------------------------------------
 #  IMPORT
 # ----------------------------------------------------------------------------------------------------------------------
+import numpy as np
+
+
 def compute_bus_conditions(bus, state, t_idx, delta_t): 
     """Computes the conditions of the bus based on the response of the battery modules
     
@@ -43,7 +46,7 @@ def compute_bus_conditions(bus, state, t_idx, delta_t):
     
     if t_idx != state.numerics.number_of_control_points-1:  
         # Handle fully charged state
-        if state.conditions.energy.recharging and bus_conditions.state_of_charge[t_idx+1] == 1:
+        if state.conditions.energy.recharging and np.float16(bus_conditions.state_of_charge[t_idx+1]) == 1:
             bus_conditions.charging_current[t_idx+1] = 0
             bus_conditions.power_draw[t_idx+1]       = 0
             bus_conditions.current_draw[t_idx+1]     = 0
